adjust int overflow to long

+++ features/merchant/seller_menu/src/main/java/com/tokopedia/seller/menu/presentation/adapter/SellerMenuAdapter.kt
@@ -30,7 +30,7 @@ class SellerMenuAdapter(
         }
     }
 
-    fun showShopInfo(shopInfo: SettingShopInfoUiModel, shopScore: Int, shopAge: Int) {
+    fun showShopInfo(shopInfo: SettingShopInfoUiModel, shopScore: Long, shopAge: Long) {
         findShopInfoIndex()?.let { index ->
             val shopInfoUiModel = ShopInfoUiModel(shopInfo, shopScore, shopAge)
             updateItemAt(index, shopInfoUiModel)
