+++ features/merchant/shop_score/src/main/java/com/tokopedia/shop/score/performance/presentation/adapter/ShopPerformanceListener.kt
@@ -1,7 +1,7 @@
 package com.tokopedia.shop.score.performance.presentation.adapter
 
 interface ShopPerformanceListener {
-    fun onTooltipLevelClicked(level: Int)
+    fun onTooltipLevelClicked(level: Long)
     fun onTooltipScoreClicked()
     fun onTickerClickedToPenaltyPage()
     fun onTickerImpressionToPenaltyPage()
